# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Schautrack API",
    "version": "1.0.0",
    "summary": "Read and write your Schautrack calorie, weight, todo, and food data.",
...
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "contact": {
      "name": "Schautrack",
      "url": "https://schautrack.com"
    }
  },
  "servers": [
    {
      "url": "https://schautrack.com/api/v1",
      "description": "This instance. Every deployment serves its own URL here, so tools that read this document — Swagger UI's \"Try it out\" included — send credentials only to the host the document came from."
    }
  ],
...
  "paths": {
    "/ai/estimate": {
      "post": {
        "operationId": "estimateFromPhoto",
        "summary": "Estimate nutrition from a food photo",
        "description": "Requires the `ai:estimate` scope, which no other scope implies — every call spends the operator's AI budget, so it must be granted deliberately. The account's daily AI limit applies here exactly as it does in the app, as does a per-account rate limit matching the app's own estimate endpoint — this is not a cheaper path to the provider. Returns 404 when no AI provider is configured; `GET /me` reports whether it is, under `server.features.ai_estimate`. Honours `Idempotency-Key`, and this is the endpoint it matters most on: a retried estimate would spend the AI budget twice and burn a second slot in the account's daily cap, for a photo already analysed.",
        "tags": [
          "AI"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Optional. A key you generate once per logical operation and reuse when retrying. The first request executes and its response is stored; a retry with the same key replays that response instead of creating a second record, and carries `Idempotency-Replayed: true`. Reusing a key for a different request body is rejected with 409. Keys are remembered for 24 hours.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The estimate.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Estimate"
                }
              }
            },
            "headers": {
              "Idempotency-Replayed": {
                "description": "Present and `true` only on a replayed response — the request matched a stored Idempotency-Key and nothing new was created. Absent on the first request. This is the only way to distinguish a replay from a fresh create, since a replay repeats the original status code.",
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "The request is malformed — unparseable JSON, an unknown field, or a bad query parameter.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "No token, or the token is unknown, revoked, or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The token is valid but lacks the scope this endpoint requires. The `required_scope` field names it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such resource, or it belongs to another account.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "The image exceeds the 10 MB limit.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited, or the account's daily AI allowance is used up (type `.../problems/ai-daily-limit`).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server-side failure.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [
              "ai:estimate"
            ]
          }
        ]
      }
    },
    "/barcode/{code}": {
      "get": {
        "operationId": "lookupBarcode",
        "summary": "Look up a product by barcode",
        "description": "Resolves an EAN-8, UPC-A, or EAN-13 barcode via OpenFoodFacts. Rate limited per account at the same rate as the app's own scanner, since it queries the same third-party database. Returns 404 when barcode lookup is disabled on the server.",
        "tags": [
          "Saved foods"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "description": "An 8-13 digit barcode. The GS1 check digit is verified.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The product.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarcodeProduct"
                }
              }
            }
          },
          "400": {
            "description": "The request is malformed — unparseable JSON, an unknown field, or a bad query parameter.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "No token, or the token is unknown, revoked, or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The token is valid but lacks the scope this endpoint requires. The `required_scope` field names it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such resource, or it belongs to another account.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "The request is well-formed but its values are rejected. `invalid_params` lists the offending fields.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. The `Retry-After` header gives the number of seconds until the window reopens.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server-side failure.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "504": {
            "description": "The food database did not answer in time.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [
              "foods:read"
            ]
          }
        ]
      }
    },
    "/docs": {
      "get": {
        "operationId": "getDocs",
        "summary": "This document, rendered for humans",
        "description": "The same description as `/openapi.json`, rendered as a self-contained HTML reference page. Needs no token.",
        "tags": [
          "Meta"
        ],
        "responses": {
          "200": {
            "description": "The API reference page.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {}
        ]
      }
    },
...
    "/entries/{id}/components": {
      "post": {
        "operationId": "addComponent",
        "summary": "Add a component to an entry",
        "description": "The entry's calories and macros become the sum of its components: grams / 100 × value per 100 g, rounded to whole units. A value that is unknown (`null`) is left out of the sum, and a macro no component knows stays `null`. Adding a component to a direct entry replaces its stored values with that sum. At most 50 components per entry.",
        "tags": [
          "Entries"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The resource's identifier.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Optional. A key you generate once per logical operation and reuse when retrying. The first request executes and its response is stored; a retry with the same key replays that response instead of creating a second record, and carries `Idempotency-Replayed: true`. Reusing a key for a different request body is rejected with 409. Keys are remembered for 24 hours.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComponentInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The updated entry, with its new totals.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entry"
                }
              }
            },
            "headers": {
              "Idempotency-Replayed": {
                "description": "Present and `true` only on a replayed response — the request matched a stored Idempotency-Key and nothing new was created. Absent on the first request. This is the only way to distinguish a replay from a fresh create, since a replay repeats the original status code.",
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "The request is malformed — unparseable JSON, an unknown field, or a bad query parameter.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "No token, or the token is unknown, revoked, or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The token is valid but lacks the scope this endpoint requires. The `required_scope` field names it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No such resource, or it belongs to another account.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request collides with existing state — a duplicate name, a limit already reached, or a feature not enabled.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "The request body exceeds the 1 MB limit. Attached to every operation that takes a body — the cap is enforced in the decoder, not per endpoint.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "The request is well-formed but its values are rejected. `invalid_params` lists the offending fields.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. The `Retry-After` header gives the number of seconds until the window reopens.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server-side failure.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [
              "entries:write"
            ]
          }
        ]
      }
    },
...
    "/foods": {
      "post": {
        "operationId": "createFood",
        "summary": "Create a food of your own",
        "description": "Adds a food with values from, say, a package label. Only on request: estimates belong in an ad-hoc component, not in the catalog.",
        "tags": [
          "Foods"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Optional. A key you generate once per logical operation and reuse when retrying. The first request executes and its response is stored; a retry with the same key replays that response instead of creating a second record, and carries `Idempotency-Replayed: true`. Reusing a key for a different request body is rejected with 409. Keys are remembered for 24 hours.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoodInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created food.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Food"
                }
              }
            },
            "headers": {
              "Idempotency-Replayed": {
                "description": "Present and `true` only on a replayed response — the request matched a stored Idempotency-Key and nothing new was created. Absent on the first request. This is the only way to distinguish a replay from a fresh create, since a replay repeats the original status code.",
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "The request is malformed — unparseable JSON, an unknown field, or a bad query parameter.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "No token, or the token is unknown, revoked, or expired.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The token is valid but lacks the scope this endpoint requires. The `required_scope` field names it.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request collides with existing state — a duplicate name, a limit already reached, or a feature not enabled.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "The request body exceeds the 1 MB limit. Attached to every operation that takes a body — the cap is enforced in the decoder, not per endpoint.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "The request is well-formed but its values are rejected. `invalid_params` lists the offending fields.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests. The `Retry-After` header gives the number of seconds until the window reopens.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server-side failure.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [
              "foods:write"
            ]
          }
        ]
      }
    },
...
    "/openapi.json": {
      "get": {
        "operationId": "getOpenAPI",
        "summary": "This document",
        "description": "Returns this OpenAPI description. Needs no token.",
        "tags": [
          "Meta"
        ],
        "responses": {
          "200": {
            "description": "The OpenAPI 3.1 document.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "security": [
          {}
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BarcodeProduct": {
        "type": "object",
        "description": "The product as resolved from OpenFoodFacts, or a not-found result.",
        "additionalProperties": true
      },
...
      "Estimate": {
        "type": "object",
        "description": "The model's nutrition estimate. Shape follows the app's AI response.",
        "additionalProperties": true
      },
      "EstimateInput": {
        "type": "object",
        "description": "A food photo to estimate.",
        "properties": {
          "context": {
            "type": "string",
            "description": "Optional hint, e.g. \"a large bowl\"."
          },
          "image": {
            "type": "string",
            "description": "The photo as a `data:image/...;base64,...` URI. Maximum 10 MB."
          }
        },
        "required": [
          "image"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "stk_\u003c43 base64url characters\u003e",
        "description": "A personal access token from Settings → API tokens. Send it as `Authorization: Bearer stk_…`."
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}